Fix flakiness of integration tests
